+++ arc/__init__.py
@@ -5,7 +5,7 @@ from arc.cli import CLI
 from arc._command import Command
 from arc.command_builders import command, namespace
 from arc.config import config, configure
-from arc.errors import ExecutionError, ValidationError, ConversionError
+from arc.errors import ExecutionError, ValidationError, ConversionError, Exit
 from arc.result import Err, Ok, Result
 from arc.context import Context
 from arc.types import *